COMP: SlicerQt - remove warning - invalid conversion from int to unsigned long.
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=12116
git-svn-id: http://svn.slicer.org/Slicer4/trunk@12116 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Libs/qVTKWidgets/Testing/qVTKObjectTest.cxx
@@ -36,16 +36,16 @@ bool qVTKObjectTest::test()
   // should do nothing but shouldn't fail neither
   qDebug() << "The following can generate error message.";
   qDebug() << "Disconnect:";
-  this->qvtkDisconnect(0, -1, this, SLOT(onVTKObjectModifiedPublic()));
+  this->qvtkDisconnect(0, static_cast<unsigned long>(-1), this, SLOT(onVTKObjectModifiedPublic()));
   qDebug() << "Connect:";
-  QString connection = this->qvtkConnect(0, -1, this, SLOT(onVTKObjectModifiedPublic()));
+  QString connection = this->qvtkConnect(0, static_cast<unsigned long>(-1), this, SLOT(onVTKObjectModifiedPublic()));
   if (!connection.isEmpty())
     {
     qDebug() << "qVTKObject::qvtkConnect() failed: "<< connection;
     return false;
     }
   qDebug() << "Reconnect:";
-  connection = this->qvtkReconnect(0, 0, -1, this, SLOT(onVTKObjectModifiedPublic()));
+  connection = this->qvtkReconnect(0, 0, static_cast<unsigned long>(-1), this, SLOT(onVTKObjectModifiedPublic()));
   if (!connection.isEmpty())
     {
     qDebug() << "qVTKObject::qvtkReconnect() failed: "<< connection;
